+++ adb-20211201/alibabacloud_adb20211201/models.py
@@ -7109,11 +7109,15 @@ class DescribeApsResourceGroupsRequest(TeaModel):
     def __init__(
         self,
         dbcluster_id: str = None,
+        region_id: str = None,
+        workload_id: str = None,
     ):
         # The ID of the AnalyticDB for MySQL Data Lakehouse Edition (V3.0) cluster.
         # 
         # >  You can call the [DescribeDBClusters](~~612397~~) operation to query the IDs of all AnalyticDB for MySQL Data Lakehouse Edition (V3.0) clusters within a region.
         self.dbcluster_id = dbcluster_id
+        self.region_id = region_id
+        self.workload_id = workload_id
 
     def validate(self):
         pass
@@ -7126,12 +7130,20 @@ class DescribeApsResourceGroupsRequest(TeaModel):
         result = dict()
         if self.dbcluster_id is not None:
             result['DBClusterId'] = self.dbcluster_id
+        if self.region_id is not None:
+            result['RegionId'] = self.region_id
+        if self.workload_id is not None:
+            result['WorkloadId'] = self.workload_id
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
         if m.get('DBClusterId') is not None:
             self.dbcluster_id = m.get('DBClusterId')
+        if m.get('RegionId') is not None:
+            self.region_id = m.get('RegionId')
+        if m.get('WorkloadId') is not None:
+            self.workload_id = m.get('WorkloadId')
         return self